+++ src/server/mod.rs
@@ -2,10 +2,14 @@
 
 use std::net::{SocketAddr, ToSocketAddrs};
 use std::sync::{Arc, Mutex};
+use std::time::Duration;
+use std::fmt::Debug;
 
+use r2d2;
+use r2d2_redis::RedisConnectionManager;
 use iron::Iron;
 use router::Router;
-use redis::{IntoConnectionInfo, ConnectionInfo};
+use redis::IntoConnectionInfo;
 
 mod handlers;
 
@@ -14,6 +18,7 @@ use ::api;
 use ::sensors;
 use ::modifiers;
 use ::errors::SpaceapiServerError;
+use ::types::RedisPool;
 
 
 /// A Space API server instance.
@@ -26,7 +31,7 @@ use ::errors::SpaceapiServerError;
 pub struct SpaceapiServer {
     socket_addr: SocketAddr,
     status: api::Status,
-    redis_connection_info: ConnectionInfo,
+    redis_pool: RedisPool,
     sensor_specs: sensors::SafeSensorSpecs,
     status_modifiers: Vec<Box<modifiers::StatusModifier>>,
 }
@@ -38,14 +43,38 @@ impl SpaceapiServer {
                      redis_connection_info: T,
                      status_modifiers: Vec<Box<modifiers::StatusModifier>>)
         -> Result<SpaceapiServer, SpaceapiServerError>
-        where U: ToSocketAddrs, T: IntoConnectionInfo {
+        where U: ToSocketAddrs, T: IntoConnectionInfo + Debug {
+            // Get socket addr
             let mut socket_addr_iter = try!(socket_addr.to_socket_addrs());
             let socket_addr = try!(socket_addr_iter.next()
                                    .ok_or(SpaceapiServerError::Message("Invalid socket address".into())));
+
+            // Log some useful debug information
+            debug!("Redis connection info: {:?}", &redis_connection_info);
+
+            // Create redis pool
+            let redis_config = r2d2::Config::builder()
+                // Provide up to 6 connections in connection pool
+                .pool_size(6)
+                // At least 1 connection must be active
+                .min_idle(Some(2))
+                // Initialize connection pool lazily. This allows the SpaceAPI
+                // server to work even without a database connection.
+                .initialization_fail_fast(false)
+                // Try to get a connection for max 1 second
+                .connection_timeout(Duration::from_secs(1))
+                // Don't log errors directly.
+                // They can get quite verbose, and we're already catching and
+                // logging the corresponding results anyways.
+                .error_handler(Box::new(r2d2::NopErrorHandler))
+                .build();
+            let redis_manager = try!(RedisConnectionManager::new(redis_connection_info));
+            let pool = try!(r2d2::Pool::new(redis_config, redis_manager));
+
             Ok(SpaceapiServer {
                 socket_addr: socket_addr,
                 status: status,
-                redis_connection_info: try!(redis_connection_info.into_connection_info()),
+                redis_pool: pool,
                 sensor_specs: Arc::new(Mutex::new(vec![])),
                 status_modifiers: status_modifiers,
             })
@@ -56,11 +85,11 @@ impl SpaceapiServer {
         let mut router = Router::new();
 
         router.get("/", handlers::ReadHandler::new(
-            self.status.clone(), self.redis_connection_info.clone(),
+            self.status.clone(), self.redis_pool.clone(),
             self.sensor_specs.clone(), self.status_modifiers));
 
         router.put("/sensors/:sensor/", handlers::UpdateHandler::new(
-            self.redis_connection_info.clone(), self.sensor_specs.clone()));
+            self.redis_pool.clone(), self.sensor_specs.clone()));
 
         router
     }
@@ -71,9 +100,6 @@ impl SpaceapiServer {
     /// http://ironframework.io/doc/hyper/server/struct.Listening.html
     /// for more information.
     pub fn serve(self) -> ::HttpResult<::Listening> {
-        // Log some useful debug information
-        debug!("Redis connection info: {:?}", &self.redis_connection_info);
-
         // Launch server process
         let socket_addr = self.socket_addr;
         let router = self.route();